Server: 부동 소수점 오차를 고려해 Update를 정수 기반으로 변경
기존에는 float를 사용해서 초단위로 계산했는데, ushort와 uint를 사용해 정수 기반으로 변경하였음.

+++ Server/MiniServerProject/Application/Stages/StageService.cs
@@ -295,7 +295,7 @@ namespace MiniServerProject.Application.Stages
             return response;
         }
 
-        public async Task<VerifyStageBattleResponse> VerifyBattleAsync(ulong userId, string requestId, string stageId, List<float> updateIntervals, CancellationToken ct, int testDelayMs = 0)
+        public async Task<VerifyStageBattleResponse> VerifyBattleAsync(ulong userId, string requestId, string stageId, List<ushort> updateIntervals, CancellationToken ct, int testDelayMs = 0)
         {
             var user = await _db.Users.AsNoTracking().FirstOrDefaultAsync(x => x.UserId == userId, ct)
                        ?? throw new DomainException(ErrorType.UserNotFound);
